requirements.txt
kafka-python==2.0.2
langchain==0.0.350
chromadb==0.4.18
openai==1.3.7
tiktoken==0.5.2
streamlit==1.28.2
pandas==2.1.4
numpy==1.24.3
faker==20.1.0
python-dotenv==1.0.0
uuid==1.30
json5==0.9.14

